feat: move addexpense page content to modal
